converts DkanStreamWrapperTest to KTB
